added reprojection method to the symmetry operator -- no need for orthogonalization and linear dependency thresholds
