+++ src/apps/chem/pointgroupsymmetry.h
@@ -46,16 +46,20 @@
 #ifndef SRC_APPS_CHEM_POINTGROUPSYMMETRY_H_
 #define SRC_APPS_CHEM_POINTGROUPSYMMETRY_H_
 
-#include <madness/mra/mra.h>
-#include <madness/mra/vmra.h>
-#include <madness/mra/functypedefs.h>
+#include <madness/world/MADworld.h>
 #include <chem/pointgroupoperator.h>
 
 using namespace madness;
 
 namespace madness {
 
+	template<typename T, std::size_t NDIM>
+	class Function;
+}
 
+
+
+namespace madness {
 struct charactertable {
 	typedef std::vector<int> characterlineT;
 
@@ -64,8 +68,19 @@ struct charactertable {
 	std::map<std::string,characterlineT> irreps_;
 	std::vector<pg_operator> operators_;		///< symmetry operators
 	std::vector<std::string> mullikan_;			///< Mullikan symbols of the irreps
+	bool is_abelian=true;						///< currently always true
 
 };
+
+/*!
+  \file projector_irrep.h
+  \brief Defines and implements the symmetry projector class
+  \ingroup chem
+  \addtogroup chem
+
+  \par Introduction
+
+*/
 
 class projector_irrep {
 
@@ -265,6 +280,15 @@ public:
 		}
 	}
 
+	/// (re-)project the argument on the given irreps
+
+	/// @param[in]	vrhs	the vector of functions to be projected on the irreps given by irreps
+	/// @param[in]	irreps	the irreps in order of the functions
+	/// @return		vrhs projected on the irreps
+	template<typename T, std::size_t NDIM>
+	std::vector<Function<T,NDIM> > project_on_irreps(const std::vector<Function<T,NDIM> >& vhrs,
+			const std::vector<std::string>& irreps) const;
+
 	/// reduce a reducible representation
 
 	/// @param[in]	reps	vector or irrep or reduc. reps whose product will be reduced
@@ -392,203 +416,9 @@ private:
 	std::vector<Function<T,NDIM> > apply_symmetry_operators(
 			const std::vector<Function<T,NDIM> >& vrhs,
 			Function<T,NDIM> metric,
-			std::vector<std::string>& sirreps) const {
+			std::vector<std::string>& sirreps) const;
 
-		World& world=vrhs.begin()->world();
-
-		// fast return
-		if (vrhs.size()==0) return std::vector<Function<T,NDIM> >();
-		if (table_.schoenflies_=="C1") {return copy(world,vrhs);}
-
-		// loop over all symmetry operations of this point group and apply them
-		// to the input vector; dimension: (nop, nmo)
-		std::vector<std::vector<Function<T,NDIM> > > opvrhs(table_.operators_.size());
-		for (int i=0; i<table_.operators_.size(); ++i) {
-			opvrhs[i]=table_.operators_[i](vrhs,false);
-		}
-
-		// get all irreps to work on
-		std::vector<std::string> all_irreps;
-		if (irrep_=="all") {
-			all_irreps=table_.mullikan_;
-		} else {
-			all_irreps.push_back(irrep_);
-		}
-
-		// accumulate the linearly combined results here; dimension: (nirrep, nmo)
-		std::vector<std::vector<Function<T,NDIM> > > lc_op_vrhs(all_irreps.size());
-		for (auto& v : lc_op_vrhs) {
-			v=zero_functions_compressed<T,NDIM>(world,vrhs.size(),false);
-		}
-
-		// need to fence here before the linear combination with the character weights
-		world.gop.fence();
-
-		// loop over all requested irreps
-		for (int j=0; j<all_irreps.size(); ++j) {
-			std::string& irrep=all_irreps[j];
-
-			// get the characters of this irrep
-			const charactertable::characterlineT cline=table_.irreps_.find(irrep)->second;
-			MADNESS_ASSERT(table_.operators_.size()==cline.size());
-
-
-			// do the linear combination; loop over all operators
-			// lc_op_vhrs[jrrep] = \sum_iop (character[iop]/table_.order_)*opvhrs[iop];
-			for (int iop=0; iop<cline.size(); ++iop) {
-				double character=double(cline[iop]);
-				gaxpy(world,1.0,lc_op_vrhs[j],character/table_.order_,opvrhs[iop],false);
-			}
-		}
-
-		// need to fence here before the orthogonalization step
-		world.gop.fence();
-		for (auto& v : lc_op_vrhs) truncate(world,v,0.0,false);
-		world.gop.fence();
-
-		// now we have a linearly dependent set of symmetrized functions
-		std::vector<Function<T,NDIM> > result1;
-		std::vector<std::string> sirreps1;
-		std::vector<int> ipiv1;
-
-		// loop over all irreps
-		int iresult=0;	// counter for the result vector ordering
-		for (int j=0; j<all_irreps.size(); ++j) {
-
-			if (verbosity_>1) print("working on irrep ",all_irreps[j]);
-			// compute overlap, include metric if necessary
-			Tensor<double> ovlp;
-			if (metric.is_initialized()) ovlp=matrix_inner(world,lc_op_vrhs[j],metric*lc_op_vrhs[j]);
-			else ovlp=matrix_inner(world,lc_op_vrhs[j],lc_op_vrhs[j]);
-
-//			for (int i=0; i<ovlp.dim(0); ++i) ovlp(i,i)+=1.e-6;
-
-			if (verbosity_>1) {
-				print("ovlp");
-				print(ovlp);
-			}
-
-			// check for zero rank
-			Tensor<double> evec,eval;
-			syev(ovlp,evec,eval);
-
-			if (eval.sum()>lindep_) {
-
-				if (verbosity_>1) print("eigenvalues",eval);
-
-				// compute rank-revealing Cholesky, discard linear dependent vectors
-				Tensor<int> piv;
-				int rank;
-				rr_cholesky(ovlp,lindep_,piv,rank);
-				Tensor<double> shrunk=copy(ovlp(Slice(0,rank-1,1),Slice(0,rank-1,1)));
-				Tensor<double> inv=inverse(shrunk);
-
-				if (verbosity_>2) {
-					print("cholesky");
-					print(ovlp);
-					print("inverse(cholesky)");
-					print(inv);
-					print("orthogonality");
-					print(inner(shrunk,shrunk,0,0));
-				}
-				if (verbosity_>1) {
-					print("pivot elements");
-					print(piv);
-				}
-
-				// collect non-vanishing elements
-				std::vector<Function<T,NDIM> > a1shrunk;
-				for (int i=0; i<rank; ++i) a1shrunk.push_back(lc_op_vrhs[j][piv(i)]);
-
-				// orthonormalize vectors within this irrep
-				if (orthonormalize_irreps_) a1shrunk=transform(world,a1shrunk,inv);
-
-				// collect all result functions and their irreps
-				for (int i=0; i<a1shrunk.size(); ++i) {
-					result1.push_back(a1shrunk[i]);
-					sirreps1.push_back(all_irreps[j]);
-					ipiv1.push_back(piv[i]);
-				}
-			}
-		}
-
-		// check closure
-		if (vrhs.size()>1 and (sirreps1.size()>vrhs.size())) {
-
-			print("vrhs.size()            ",vrhs.size());
-			print("raw sirreps            ",sirreps1);
-			print("raw mapping in ipiv1   ",ipiv1);
-
-			MADNESS_EXCEPTION("\n\nfunction arguments in apply_symmetry_operators are not closed\n\n",1);
-		}
-
-		if (verbosity_>1) {
-			print("raw sirreps            ",sirreps1);
-			print("raw mapping in ipiv1   ",ipiv1);
-		}
-
-		// resorting section, does not apply if number of function increases
-
-		// if linear dependencies occur, the mapping of input and output functions
-		// might not be unique, i.e. the mapping might look like ipiv={0,1,1,3,4},
-		// with double entries (1) and missing entries (2).
-		// Imagine a He dimer with the bonding and antibonding orbitals both being mapped
-		// to the left atomic orbital, just by numerical noise.
-		// Find double and missing entries and replace the double with the missing ones
-
-		std::vector<Function<T,NDIM> > result=zero_functions<T,NDIM>(world,vrhs.size());
-		sirreps=std::vector<std::string>(vrhs.size(),"null");
-
-		if (vrhs.size()==1) {
-			result=result1;
-			sirreps=sirreps1;
-
-		} else {
-			// find all double elements
-			std::vector<int> ipiv2(ipiv1),double_elements;
-			std::sort(ipiv2.begin(), ipiv2.end());
-			for (int i=1; i<ipiv2.size(); ++i) {
-				if (ipiv2[i]==ipiv2[i-1]) double_elements.push_back(ipiv2[i]);
-			}
-
-			if (verbosity_>1) print("double elements  ", double_elements);
-
-			// if there are double elements we have to recompute the overlap matrix
-			// to safely determine the mapping
-			if (double_elements.size()>0) {
-				Tensor<double> ovlp1=matrix_inner(world,result1,vrhs);
-
-				long index[2];
-				// find the largest overlap matrix element in each column
-				for (int i=0; i<ovlp1.dim(1); ++i) {
-
-					ovlp1(_,i).absmax(index);
-					if (verbosity_>2) {
-						print("result overlap matrix");
-						print(ovlp1);
-						print("index",index[0]);
-					}
-					ipiv1[i]=index[0];
-					// remove this row from the tensor
-					ovlp1(index[0],_)=0.0;
-				}
-			}
-			if (verbosity_>1) print("final ipiv1",ipiv1);
-
-			for (int i=0; i<ipiv1.size(); ++i) {
-				result[ipiv1[i]]=result1[i];
-				sirreps[ipiv1[i]]=sirreps1[i];
-			}
-		}	// end of resorting/mapping section
-
-		if (not keep_ordering_) result=sort_to_irreps(result,sirreps);
-		if (verbosity_>0) print("final function irreps: ",sirreps);
-
-		return result;
-
-	}
-
-	/// given a vector of functions and their irreps, extract only the ones of the requested irrep
+	/// sort the functions according to their irreps
 	template<typename T, std::size_t NDIM>
 	std::vector<Function<T,NDIM> > sort_to_irreps(std::vector<Function<T,NDIM> >& vrhs,
 			std::vector<std::string>& sirreps) const {
